@@ -4,7 +4,7 @@
 
 from __future__ import annotations
 
-from typing import Iterable
+from collections.abc import Iterable
 
 from ._data import RetryHook, RetryHookFactory
 from ._logging import LoggingOnRetryHook
