@@ -4,9 +4,10 @@
 
 from __future__ import annotations
 
+from collections.abc import Callable
 from contextlib import AbstractContextManager
 from dataclasses import dataclass
-from typing import Callable, Protocol
+from typing import Protocol
 
 
 def guess_name(obj: object) -> str:
